+++ breaditnow-auth/src/main/java/com/breaditnow/auth/BreaditnowAuthApplication.java
@@ -1,4 +1,4 @@
-package com.breaditnow;
+package com.breaditnow.auth;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
