+++ breaditnow-customer-api/build.gradle
@@ -15,6 +15,7 @@ dependencies {
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'com.mysql:mysql-connector-j'
 
     implementation project(":breaditnow-mysql")
 }
