+++ breaditnow-mysql/build.gradle
@@ -11,10 +11,10 @@ repositories {
 
 dependencies {
     developmentOnly("org.springframework.boot:spring-boot-docker-compose")
-    runtimeOnly 'com.mysql:mysql-connector-j'
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'com.mysql:mysql-connector-j'
 }
 
 test {
