+++ breaditnow-mysql/src/main/java/com/breaditnow/mysql/repository/BreaditnowMysqlApplication.java
@@ -1,11 +1,11 @@
-package com.breaditnow;
+package com.breaditnow.mysql.repository;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class BreaditnowOwnerApplication {
+public class BreaditnowMysqlApplication {
     public static void main(String[] args) {
-        SpringApplication.run(BreaditnowOwnerApplication.class, args);
+        SpringApplication.run(BreaditnowMysqlApplication.class, args);
     }
 }
