+++ breaditnow-owner-api/src/main/java/com/breaditnow/owner/BreaditnowOwnerApplication.java
@@ -1,11 +1,11 @@
-package com.breaditnow.mysql;
+package com.breaditnow.owner;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class BreaditnowMysqlApplication {
+public class BreaditnowOwnerApplication {
     public static void main(String[] args) {
-        SpringApplication.run(BreaditnowMysqlApplication.class, args);
+        SpringApplication.run(BreaditnowOwnerApplication.class, args);
     }
 }
