+++ breaditnow-owner-api/src/main/resources/application.yml
@@ -4,3 +4,6 @@ server:
 spring:
   application:
     name: breaditnow-owner-api
+
+  profiles:
+    active: local
